refactoring: removed unused files
