+++ .env.example
@@ -1,5 +1,7 @@
 PORT=4000
 
+DATABASE_URL="postgres://myuser:mypassword@localhost:5433/mydb"
+
 CRYPT_SALT=10
 JWT_SECRET_KEY=secret123123
 JWT_SECRET_REFRESH_KEY=secret123123
