--- db-store/const.ts
@@ -1,4 +1,0 @@
-export const USER_ID = '5d32bcc0-dd5c-11ee-8513-356525c30cc5';
-export const TRACK_ID = '7fbbae03-dd5c-11ee-a686-ed5a6659a3e1';
-export const ALBUM_ID = '201702a0-dd5d-11ee-82b9-f35a2947a09a';
-export const ARTIST_ID = '80458200-dd5d-11ee-a97d-c5cef3c97348';
