+++ docker-compose.yml
@@ -1,0 +1,16 @@
+version: '3.8'
+
+services:
+    postgres:
+        image: postgres:13.5
+        restart: always
+        environment:
+            - POSTGRES_USER=myuser
+            - POSTGRES_PASSWORD=mypassword
+            - POSTGRES_DB=mydb
+        volumes:
+            - postgres:/var/lib/postgresql/data
+        ports:
+            - '5433:5432'
+volumes:
+    postgres:
