+++ src/utils/getSwaggerDoc.ts
@@ -3,7 +3,7 @@ import { join } from 'path';
 import { parse } from 'yaml';
 
 export const getSwaggerDocument = () => {
-    const pathToApiYaml = join(__dirname, '../../../doc/api.yaml');
+    const pathToApiYaml = join(__dirname, '../../doc/api.yaml');
     const file = readFileSync(pathToApiYaml, 'utf8');
     return parse(file);
 };
